requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==23.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.9.0
tzdata==2021.5
